+++ googledataprocauthenticator/controllerwidget/controllerwidget.py
@@ -15,7 +15,7 @@
 import ipyvuetify as v
 from sparkmagic.controllerwidget.abstractmenuwidget import AbstractMenuWidget
 from googledataprocauthenticator.controllerwidget.addendpointwidget import AddEndpointWidget
-from googledataprocauthenticator.magics.dataprocmagics import SerializableEndpoint
+from googledataprocauthenticator.utils.serializeableendpoint import SerializableEndpoint
 from googledataprocauthenticator.controllerwidget.createsessionwidget import CreateSessionWidget
 from sparkmagic.controllerwidget.manageendpointwidget import ManageEndpointWidget
 from sparkmagic.controllerwidget.managesessionwidget import ManageSessionWidget
